refactor(drcp-format): Factor out a lineparser in the reader
